Added a new Announcement version 1 format which:
1. Moves 98% of the effort to the dataset creation phase
2. Encrypts the merkle proof and last section of the announcement using
the CryptoCycle state, thus making it impossible to decompress the
announcement without doing as much effort as is needed to validate it.
3. SoftNonce is constrained based on difficulty such that only about 512
announcements will be won per dataset.
4. Datasets are created twice, in order that merkle root corruption will
cost at least half as much as honest dataset re-creation.

The net effect of this is one on one core of a 2015 2.5Ghz Intel i7, one
can mine approximately 1 announcement per millisecond at the minimum
difficulty. As one needs to re-create the dataset for every 512
announcements, performance increases as difficulty rises so at 4096
difficulty, announcements are produced at a rate of approximately 1
announcement every 26ms.